refactor(mocker): teach replay runtimes some manners
