+++ lib/mocker/src/replay/runtime/mod.rs
@@ -3,11 +3,46 @@
 
 use crate::common::protocols::{DirectRequest, MockEngineArgs};
 use crate::replay::TraceSimulationReport;
+use std::collections::VecDeque;
 
 pub(crate) mod core;
 pub(crate) mod events;
 pub(crate) mod multi;
 pub(crate) mod single;
+pub(crate) mod state;
+
+fn normalize_trace_requests(
+    mut requests: Vec<DirectRequest>,
+) -> anyhow::Result<VecDeque<DirectRequest>> {
+    requests.sort_by(|left, right| {
+        let left_ts = left
+            .arrival_timestamp_ms
+            .expect("trace replay requests must have an arrival timestamp");
+        let right_ts = right
+            .arrival_timestamp_ms
+            .expect("trace replay requests must have an arrival timestamp");
+        left_ts.total_cmp(&right_ts)
+    });
+
+    let first_arrival_ms = requests
+        .first()
+        .and_then(|request| request.arrival_timestamp_ms)
+        .ok_or_else(|| anyhow::anyhow!("trace replay requires at least one timestamped request"))?;
+
+    Ok(VecDeque::from(
+        requests
+            .into_iter()
+            .map(|mut request| {
+                let arrival_timestamp_ms = request
+                    .arrival_timestamp_ms
+                    .expect("trace replay requests must have an arrival timestamp")
+                    - first_arrival_ms;
+                request.arrival_timestamp_ms = Some(arrival_timestamp_ms);
+                request
+            })
+            .collect::<Vec<_>>(),
+    ))
+}
 
 pub(crate) fn simulate_trace(
     args: MockEngineArgs,
